Add HA dashboard for heating
Added a version of dashboard showing temperatures
on a process diagram. Also, added a svg-file and
corresponding png file of the image that is used
in the picture-elements card as the backround
on top of which state labels of temperatures are
added.

Picture files contain texts in finnish language.
Other languages should be easy to produce using the
svg file as basis.